Add automated releases (#1127)
Add workflows to automate releases if the Java instrumentation version
has changed since the last release.

If it has, then that version is used for the release.

Only changes to the following line of code are considered:


https://github.com/grafana/grafana-opentelemetry-java/blob/fd221d19fb50ad1b5096c7a518fb33623f6bbd19/build.gradle#L6

Any other changes will not trigger a release.

---

Based on the implementation from the following PRs:

- https://github.com/grafana/docker-otel-lgtm/pull/868
- https://github.com/grafana/docker-otel-lgtm/pull/913
- https://github.com/grafana/docker-otel-lgtm/pull/917

## TODO

- [x] ~~Add secrets to Vault~~
- [x] ~~Update GitHub token broker configuration~~
- [x] ~~Verify generating GitHub tokens~~
[Test](https://github.com/grafana/grafana-opentelemetry-java/actions/runs/20234659806/job/58086033304#step:3:7)
- [x] ~~Update
[`RELEASING.md`](https://github.com/grafana/grafana-opentelemetry-java/blob/main/RELEASING.md)~~